+++ example/lib/domain/usecases/create_user.dart
@@ -1,8 +1,9 @@
+import 'package:dartz/dartz.dart';
 import 'package:fakestore_fase3_mandrade/fakestore_fase3_mandrade.dart';
 
 class CreateUser {
   final UserRepository _repo;
   const CreateUser(this._repo);
 
-  Future<UserModel> call(CreateUserInput input) => _repo.createUser(input);
+  Future<Either<Failure, UserModel>> call(CreateUserInput input) => _repo.createUser(input);
 }