+++ example/lib/domain/usecases/get_all_carts.dart
@@ -1,8 +1,9 @@
+import 'package:dartz/dartz.dart';
 import 'package:fakestore_fase3_mandrade/fakestore_fase3_mandrade.dart';
 
 class GetAllCarts {
   final CartRepository _repo;
   const GetAllCarts(this._repo);
 
-  Future<List<CartModel>> call({int? limit, String? sort}) => _repo.getAllCarts(limit: limit, sort: sort);
+  Future<Either<Failure, List<CartModel>>> call({int? limit, String? sort}) => _repo.getAllCarts(limit: limit, sort: sort);
 }