+++ example/lib/domain/usecases/get_all_products.dart
@@ -4,5 +4,5 @@ class GetAllProducts {
   final ProductRepository _repo;
   const GetAllProducts(this._repo);
 
-  Future<List<ProductModel>> call({int? limit, String? sort}) => _repo.getAllProducts(limit: limit, sort: sort);
+  Future<Either<Failure, List<ProductModel>>> call({int? limit, String? sort}) => _repo.getAllProducts(limit: limit, sort: sort);
 }