+++ example/lib/domain/usecases/get_carts_by_user_id.dart
@@ -1,8 +1,9 @@
+import 'package:dartz/dartz.dart';
 import 'package:fakestore_fase3_mandrade/fakestore_fase3_mandrade.dart';
 
 class GetCartsByUserId {
   final CartRepository _repo;
   const GetCartsByUserId(this._repo);
 
-  Future<List<CartModel>> call(int userId) => _repo.getCartsByUserId(userId);
+  Future<Either<Failure, List<CartModel>>> call(int userId) => _repo.getCartsByUserId(userId);
 }