+++ example/lib/domain/usecases/get_categories.dart
@@ -1,8 +1,9 @@
+import 'package:dartz/dartz.dart';
 import 'package:fakestore_fase3_mandrade/fakestore_fase3_mandrade.dart';
 
 class GetCategories {
   final ProductRepository _repo;
   const GetCategories(this._repo);
 
-  Future<List<String>> call() => _repo.getCategories();
+  Future<Either<Failure, List<String>>> call() => _repo.getCategories();
 }