+++ example/lib/domain/usecases/get_product_by_id.dart
@@ -1,8 +1,9 @@
+import 'package:dartz/dartz.dart';
 import 'package:fakestore_fase3_mandrade/fakestore_fase3_mandrade.dart';
 
 class GetProductById {
   final ProductRepository _repo;
   const GetProductById(this._repo);
 
-  Future<ProductModel> call(int id) => _repo.getProductById(id);
+  Future<Either<Failure, ProductModel>> call(int id) => _repo.getProductById(id);
 }