+++ example/lib/domain/usecases/get_products_by_category.dart
@@ -1,10 +1,11 @@
+import 'package:dartz/dartz.dart';
 import 'package:fakestore_fase3_mandrade/fakestore_fase3_mandrade.dart';
 
 class GetProductsByCategory {
   final ProductRepository _repo;
   const GetProductsByCategory(this._repo);
 
-  Future<List<ProductModel>> call(String category, {int? limit, String? sort}) {
+  Future<Either<Failure, List<ProductModel>>> call(String category, {int? limit, String? sort}) {
     return _repo.getProductsByCategory(category, limit: limit, sort: sort);
   }
 }