+++ example/lib/domain/usecases/get_user_by_id.dart
@@ -1,8 +1,9 @@
+import 'package:dartz/dartz.dart';
 import 'package:fakestore_fase3_mandrade/fakestore_fase3_mandrade.dart';
 
 class GetUserById {
   final UserRepository _repo;
   const GetUserById(this._repo);
 
-  Future<UserModel> call(int id) => _repo.getUserById(id);
+  Future<Either<Failure, UserModel>> call(int id) => _repo.getUserById(id);
 }