+++ lib/data/repositories/cart_repository_impl.dart
@@ -1,5 +1,7 @@
+import 'package:dartz/dartz.dart';
 import '../../domain/models/cart_model.dart';
 import '../../domain/repositories/cart_repository.dart';
+import '../../domain/failures/failures.dart';
 import '../datasources/fake_store_remote_datasource.dart';
 
 /// Implementación concreta del repositorio de carritos
@@ -10,7 +12,7 @@ class CartRepositoryImpl implements CartRepository {
   FakeStoreRemoteDataSource get dataSource => _ds;
 
   @override
-  Future<List<CartModel>> getAllCarts({
+  Future<Either<Failure, List<CartModel>>> getAllCarts({
     int? limit,
     String? sort,
     DateTime? startDate,
@@ -25,12 +27,12 @@ class CartRepositoryImpl implements CartRepository {
   }
 
   @override
-  Future<CartModel> getCartById(int id) async {
+  Future<Either<Failure, CartModel>> getCartById(int id) async {
     return _ds.getCartById(id);
   }
 
   @override
-  Future<List<CartModel>> getCartsByUserId(int userId) async {
+  Future<Either<Failure, List<CartModel>>> getCartsByUserId(int userId) async {
     return _ds.getCartsByUserId(userId);
   }
 }