+++ lib/data/repositories/product_repository_impl.dart
@@ -1,5 +1,7 @@
+import 'package:dartz/dartz.dart';
 import '../../domain/models/product_model.dart';
 import '../../domain/repositories/product_repository.dart';
+import '../../domain/failures/failures.dart';
 import '../datasources/fake_store_remote_datasource.dart';
 
 /// Implementación concreta del repositorio de productos
@@ -11,22 +13,22 @@ class ProductRepositoryImpl implements ProductRepository {
   FakeStoreRemoteDataSource get dataSource => _ds; // permite acceder a dispose si se creó internamente
 
   @override
-  Future<List<ProductModel>> getAllProducts({int? limit, String? sort}) async {
+  Future<Either<Failure, List<ProductModel>>> getAllProducts({int? limit, String? sort}) async {
     return _ds.getAllProducts(limit: limit, sort: sort);
   }
 
   @override
-  Future<ProductModel> getProductById(int id) async {
+  Future<Either<Failure, ProductModel>> getProductById(int id) async {
     return _ds.getProductById(id);
   }
 
   @override
-  Future<List<String>> getCategories() async {
+  Future<Either<Failure, List<String>>> getCategories() async {
     return _ds.getCategories();
   }
 
   @override
-  Future<List<ProductModel>> getProductsByCategory(
+  Future<Either<Failure, List<ProductModel>>> getProductsByCategory(
     String category, {
     int? limit,
     String? sort,