+++ lib/data/repositories/user_repository_impl.dart
@@ -1,6 +1,8 @@
+import 'package:dartz/dartz.dart';
 import '../../domain/models/user_model.dart';
 import '../../domain/models/create_user_input.dart';
 import '../../domain/repositories/user_repository.dart';
+import '../../domain/failures/failures.dart';
 import '../datasources/fake_store_remote_datasource.dart';
 
 /// Implementación concreta del repositorio de usuarios
@@ -11,22 +13,22 @@ class UserRepositoryImpl implements UserRepository {
   FakeStoreRemoteDataSource get dataSource => _ds;
 
   @override
-  Future<List<UserModel>> getAllUsers({int? limit, String? sort}) async {
+  Future<Either<Failure, List<UserModel>>> getAllUsers({int? limit, String? sort}) async {
     return _ds.getAllUsers(limit: limit, sort: sort);
   }
 
   @override
-  Future<UserModel> getUserById(int id) async {
+  Future<Either<Failure, UserModel>> getUserById(int id) async {
     return _ds.getUserById(id);
   }
 
   @override
-  Future<UserModel> createUser(CreateUserInput input) async {
+  Future<Either<Failure, UserModel>> createUser(CreateUserInput input) async {
     return _ds.createUser(input);
   }
 
   @override
-  Future<String> login(String username, String password) async {
+  Future<Either<Failure, String>> login(String username, String password) async {
     return _ds.login(username: username, password: password);
   }
 }