+++ lib/domain/repositories/cart_repository.dart
@@ -1,9 +1,11 @@
+import 'package:dartz/dartz.dart';
 import '../models/cart_model.dart';
+import '../failures/failures.dart';
 
 /// Contrato/Interface para el repositorio de carritos
 abstract class CartRepository {
   /// Obtiene todos los carritos
-  Future<List<CartModel>> getAllCarts({
+  Future<Either<Failure, List<CartModel>>> getAllCarts({
     int? limit,
     String? sort,
     DateTime? startDate,
@@ -11,8 +13,8 @@ abstract class CartRepository {
   });
 
   /// Obtiene un carrito por su ID
-  Future<CartModel> getCartById(int id);
+  Future<Either<Failure, CartModel>> getCartById(int id);
 
   /// Obtiene los carritos de un usuario específico
-  Future<List<CartModel>> getCartsByUserId(int userId);
+  Future<Either<Failure, List<CartModel>>> getCartsByUserId(int userId);
 }